# openapi.yaml
openapi: 3.0.3
info:
  title: GameVerse User Profile Service
  description: Production-ready user profile management API with blockchain addresses, KYC, and GDPR compliance
  version: 1.7.0
  contact:
    name: GameVerse Team
    email: support@gameverse.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.gameverse.com/api/v1
    description: Production server

tags:
  - name: Users
    description: User profile management
  - name: Blockchain
    description: Blockchain address management
  - name: KYC
    description: KYC status management
  - name: GDPR
    description: GDPR compliance endpoints

paths:
  /users:
    post:
      tags:
        - Users
      summary: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserInput'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

    get:
      tags:
        - Users
      summary: Get all users (admin only)
      operationId: getAllUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsersResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /users/search:
    get:
      tags:
        - Users
      summary: Search user by email
      operationId: getUserByEmail
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /users/verify/{token}:
    get:
      tags:
        - Users
      summary: Verify email address
      operationId: verifyEmail
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      operationId: getUser
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Users
      summary: Update user profile
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserInput'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

    delete:
      tags:
        - Users
      summary: Delete user (soft delete)
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /users/{id}/avatar:
    post:
      tags:
        - Users
      summary: Upload user avatar
      operationId: uploadAvatar
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
              required:
                - avatar
      responses:
        '200':
          description: Avatar uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /users/{id}/kyc:
    patch:
      tags:
        - KYC
      summary: Update KYC status (admin only)
      operationId: updateKycStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KycStatusUpdate'
      responses:
        '200':
          description: KYC status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /users/{id}/kyc/history:
    get:
      tags:
        - KYC
      summary: Get KYC history
      operationId: getKycHistory
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: KYC history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KycHistoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /users/{id}/export:
    get:
      tags:
        - GDPR
      summary: Export user data (GDPR)
      operationId: exportUserData
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: User data export
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GdprExportResponse'
            text/csv:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /users/{id}/anonymize:
    post:
      tags:
        - GDPR
      summary: Anonymize user data (right to be forgotten)
      operationId: anonymizeUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: User data anonymized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /users/{id}/addresses:
    post:
      tags:
        - Blockchain
      summary: Link blockchain address
      operationId: linkAddress
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkAddressInput'
      responses:
        '201':
          description: Address linked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockchainAddressResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

    get:
      tags:
        - Blockchain
      summary: Get user's blockchain addresses
      operationId: getAddresses
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: List of blockchain addresses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockchainAddressesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /users/{id}/addresses/{addressId}:
    get:
      tags:
        - Blockchain
      summary: Get specific blockchain address
      operationId: getAddress
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - $ref: '#/components/parameters/AddressIdParam'
      responses:
        '200':
          description: Blockchain address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockchainAddressResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Blockchain
      summary: Unlink blockchain address
      operationId: unlinkAddress
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - $ref: '#/components/parameters/AddressIdParam'
      responses:
        '200':
          description: Address unlinked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /users/{id}/addresses/{addressId}/primary:
    patch:
      tags:
        - Blockchain
      summary: Set address as primary
      operationId: setPrimaryAddress
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - $ref: '#/components/parameters/AddressIdParam'
      responses:
        '200':
          description: Primary address updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockchainAddressResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    UserIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid

    AddressIdParam:
      name: addressId
      in: path
      required: true
      schema:
        type: string
        format: uuid

    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1

    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  schemas:
    CreateUserInput:
      type: object
      required:
        - email
        - username
      properties:
        email:
          type: string
          format: email
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9_]+$'
        displayName:
          type: string
          maxLength: 100
        bio:
          type: string
          maxLength: 500

    UpdateUserInput:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[a-zA-Z0-9_]+$'
        displayName:
          type: string
          maxLength: 100
        bio:
          type: string
          maxLength: 500
        preferences:
          $ref: '#/components/schemas/UserPreferences'

    UserPreferences:
      type: object
      properties:
        notifications:
          type: object
          properties:
            email:
              type: boolean
            push:
              type: boolean
            marketing:
              type: boolean
            gameUpdates:
              type: boolean
            friendRequests:
              type: boolean
        privacy:
          type: object
          properties:
            profileVisibility:
              type: string
              enum: [public, friends, private]
            showOnlineStatus:
              type: boolean
            showGameActivity:
              type: boolean
            allowFriendRequests:
              type: boolean
        display:
          type: object
          properties:
            theme:
              type: string
              enum: [light, dark, system]
            compactMode:
              type: boolean
            animationsEnabled:
              type: boolean
        locale:
          type: string
        timezone:
          type: string

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        emailVerified:
          type: boolean
        kycStatus:
          $ref: '#/components/schemas/KycStatus'
        kycVerifiedAt:
          type: string
          format: date-time
          nullable: true
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    KycStatus:
      type: string
      enum: [none, pending, verified, rejected, expired]

    KycStatusUpdate:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/KycStatus'
        provider:
          type: string
        reference:
          type: string

    KycHistoryEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/KycStatus'
        provider:
          type: string
          nullable: true
        reference:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time

    BlockchainChain:
      type: string
      enum: [ethereum, polygon, solana, avalanche, binance, arbitrum, optimism, base]

    LinkAddressInput:
      type: object
      required:
        - chain
        - address
        - signature
        - message
      properties:
        chain:
          $ref: '#/components/schemas/BlockchainChain'
        address:
          type: string
        signature:
          type: string
        message:
          type: string
        label:
          type: string
          maxLength: 50

    BlockchainAddress:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        chain:
          $ref: '#/components/schemas/BlockchainChain'
        address:
          type: string
        isPrimary:
          type: boolean
        verifiedAt:
          type: string
          format: date-time
          nullable: true
        label:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/User'
        message:
          type: string

    PaginatedUsersResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/User'
            total:
              type: integer
            page:
              type: integer
            limit:
              type: integer
            totalPages:
              type: integer

    BlockchainAddressResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/BlockchainAddress'
        message:
          type: string

    BlockchainAddressesResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/BlockchainAddress'

    KycHistoryResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/KycHistoryEntry'

    GdprExportResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            profile:
              $ref: '#/components/schemas/User'
            blockchainAddresses:
              type: array
              items:
                $ref: '#/components/schemas/BlockchainAddress'
            kycHistory:
              type: array
              items:
                $ref: '#/components/schemas/KycHistoryEntry'
            exportedAt:
              type: string
              format: date-time
            format:
              type: string
              enum: [json, csv]

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: string

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'